services/api-gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = "Django Service"
package-mode = false
authors = ["Bobur Yusupov <yusufi.bobur@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0"
django-cors-headers = "^4.9.0"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
pytest-django = "^4.11.1"

[tool.poetry.group.dev.dependencies]
black = "^25.11.0"
flake8 = "^7.3.0"
mypy = "^1.19.0"
pytest = "^9.0.1"
drf-spectacular = "^0.29.0"
django-stubs = "^5.2.8"
djangorestframework-stubs = "^3.16.6"
types-requests = "^2.32.4.20250913"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::django.utils.deprecation.RemovedInDjango70Warning",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | migrations
  | __pycache__
  | node_modules
  | tests
)/
'''
